app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Jason Johnson <jasojohnson@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
aiohttp = "^3.10.5"
aiofiles = "^24.1.0"
xmltodict = "^0.13.0"
azure-functions = "^1.20.0"
pymongo = "^4.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
